Revert "this version enables floating points as inputs."
This reverts commit a7ea7807f9827632e6c8406bcd0285d0c5403783.

+++ AppDev _Act1/Program.cs
@@ -10,14 +10,14 @@ namespace AppDev_Act1
     {
         static void Main(string[] args)
         {
-            double x;
-            double y;
+            int x;
+            int y;
 
             Console.Write("Enter first number: ");
-            x = Convert.ToDouble(Console.ReadLine());
+            x = Convert.ToInt32(Console.ReadLine());
 
             Console.Write("Enter second number: ");
-            y = Convert.ToDouble(Console.ReadLine());
+            y = Convert.ToInt32(Console.ReadLine());
 
             string operation;
